requirements.txt
flask==3.0.0
openai>=1.54.0
python-dotenv==1.0.0
httpx>=0.27.0
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0
